+++ backend/src/classes/Controller/App/CharController.php
@@ -2,6 +2,7 @@
 
 namespace Neucore\Controller\App;
 
+use Neucore\Entity\Player;
 use Neucore\Factory\RepositoryFactory;
 use Slim\Http\Response;
 use Swagger\Annotations as SWG;
@@ -30,7 +31,7 @@ class CharController
      * @SWG\Get(
      *     path="/app/v1/main/{cid}",
      *     operationId="mainV1",
-     *     summary="Returns the main character of the player account to which the character ID belongs.",
+     *     summary="Return the main character of the player account to which the character ID belongs.",
      *     description="Needs role: app-chars.<br>It is possible that an account has no main character.",
      *     tags={"Application"},
      *     security={{"Bearer"={}}},
@@ -197,12 +198,47 @@ class CharController
             return $this->response->withStatus(404, 'Character not found.');
         }
 
-        $result = [];
-        foreach ($char->getPlayer()->getCharacters() as $character) {
-            $result[] = $character;
+        return $this->response->withJson($char->getPlayer()->getCharacters());
+    }
+
+    /**
+     * @SWG\Get(
+     *     path="/app/v1/player-chars/{playerId}",
+     *     operationId="playerCharactersV1",
+     *     summary="Return all characters from the player account.",
+     *     description="Needs role: app-chars.",
+     *     tags={"Application"},
+     *     security={{"Bearer"={}}},
+     *     @SWG\Parameter(
+     *         name="playerId",
+     *         in="path",
+     *         required=true,
+     *         description="Player ID.",
+     *         type="integer"
+     *     ),
+     *     @SWG\Response(
+     *         response="200",
+     *         description="All characters from the player account.",
+     *         @SWG\Schema(type="array", @SWG\Items(ref="#/definitions/Character"))
+     *     ),
+     *     @SWG\Response(
+     *         response="404",
+     *         description="Player not found."
+     *     ),
+     *     @SWG\Response(
+     *         response="403",
+     *         description="Not authorized."
+     *     )
+     * )
+     */
+    public function playerCharactersV1(string $playerId): Response
+    {
+        $player = $this->repositoryFactory->getPlayerRepository()->find((int) $playerId);
+        if ($player === null) {
+            return $this->response->withStatus(404, 'Player not found.');
         }
 
-        return $this->response->withJson($result);
+        return $this->response->withJson($player->getCharacters());
     }
 
     /**
@@ -248,5 +284,40 @@ class CharController
         }
 
         return $this->response->withJson($result);
+    }
+
+    /**
+     * @SWG\Get(
+     *     path="/app/v1/corp-players/{corporationId}",
+     *     operationId="corporationPlayersV1",
+     *     summary="Return a list of all player IDs that have a character in the corporation.",
+     *     description="Needs role: app-chars.",
+     *     tags={"Application"},
+     *     security={{"Bearer"={}}},
+     *     @SWG\Parameter(
+     *         name="corporationId",
+     *         in="path",
+     *         required=true,
+     *         description="EVE corporation ID.",
+     *         type="integer"
+     *     ),
+     *     @SWG\Response(
+     *         response="200",
+     *         description="List of players, only id and name properties are returned.",
+     *         @SWG\Schema(type="array", @SWG\Items(ref="#/definitions/Player"))
+     *     ),
+     *     @SWG\Response(
+     *         response="403",
+     *         description="Not authorized."
+     *     )
+     * )
+     */
+    public function corporationPlayersV1(string $corporationId): Response
+    {
+        $players = $this->repositoryFactory->getPlayerRepository()->findInCorporation((int) $corporationId);
+
+        return $this->response->withJson(array_map(function (Player $player) {
+            return $player->jsonSerialize(true);
+        }, $players));
     }
 }