+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# next
+
+dd mmm yyyy
+
+- Add: API endpoint for apps that returns all characters from a player account.
+- Add: API endpoint for apps that accepts a corporation ID and returns a list of all player IDs that have a 
+  character in the corporation.
+- Small improvements for UI, frontend and documentation.
+
 
 ## 1.2.1